requirements.txt
aioredis==1.3.1
altgraph==0.17
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.12.5
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
constantly==15.1.0
cryptography==3.4.7
Cython==0.29.21
daphne==2.5.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.1
django-allauth==0.44.0
django-cart==1.0.4
django-celery==3.3.1
django-countries==7.1
django-filter==2.4.0
django-import-export==2.4.0
django-money==1.3.1
django-multiselectfield==0.1.12
django-next-url-mixin==0.1.0
django-ordered-model==3.4.3
-e git://github.com/cypreess/django-plans.git@b60e0748e17c2820c92b6d9240e6fc8f32092463#egg=django_plans
django-rest-auth==0.9.5
django-sequences==2.6
djangorestframework==3.12.1
et-xmlfile==1.0.1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jdcal==1.4.1
kombu==3.0.37
macholib==1.14
MarkupPy==1.14
msgpack==1.0.0
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.2.0
plyer==1.4.3
prompt-toolkit==3.0.8
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
PyJWT==2.1.0
pyobjus==1.1.0
pyOpenSSL==19.1.0
python-stdnum==1.16
python3-openid==3.2.0
pytz==2020.1
pyuca==1.2
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
suds-jurko==0.6
tablib==2.0.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.4
vatnumber==1.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.1.2
